refactor most of temporary drawer tests into slidable drawer tests
